requirements.txt
playwright==1.39.0
playwright-stealth==1.0.6
openpyxl==3.1.2
pandas
google-cloud-storage
google-cloud-aiplatform
pillow